LPD-26258 Move getRootDir() to constants
Even though it is a function it is really a constant for the whole lifetime
of the node-scripts process.

We don't turn it into a real constant because that would make computing it
in advance necessary and, since it is expensive, it's better to delay it by
providing it as a function.